feat : BroadcastReceiver 위치 MainActivity로 변경, Timer 작동 안되던 에러 해결
- receiver를 fragment에 두니 여러개 생성되는 이슈 발생 -> MainActivity로 옮겨서 한번만 생성되도록 변경
- Timer가 background에서 시간이 오래되면 멈추는 이슈 발생 -> onResume()에서 startService() 한번 더 호출해서 onStartCommand() 함수 실행되도록 구현
- Service가 종료되어도 timer의 while문은 종료 안됨 -> 변수 생성해서 게임 오버시 while문 탈출하도록 구현
